requirements.txt
beautifulsoup4
lxml
openai
pandas
psycopg2-binary
python-dotenv
requests
tqdm
